+++ tic_tac_toe_flutter/lib/game/game.dart
@@ -1,0 +1,2 @@
+export './board.dart';
+export './player.dart';
