+++ tic_tac_toe_flutter/lib/lib.dart
@@ -1,1 +1,1 @@
-export './board/board.dart';
+export 'game/game.dart';
